requirements.txt
bcrypt==4.2.0
blinker==1.8.2
Brotli==1.1.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
cssselect2==0.7.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.54.1
html5lib==1.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
pillow==11.0.0
pycparser==2.22
pydyf==0.11.0
PyMySQL==1.1.1
PyPDF2==3.0.1
pyphen==0.16.0
python-docx==1.1.2
python-magic==0.4.27
six==1.16.0
SQLAlchemy==2.0.35
tinycss2==1.4.0
typing_extensions==4.12.2
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.4
zopfli==0.2.3.post1
